Filtern ist jetzt möglich

+++ src/main/java/de/mushroomfinder/controller/CommentsAddController.java
@@ -14,6 +14,7 @@ import javax.websocket.server.PathParam;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -49,18 +50,22 @@ public class CommentsAddController {
 	public ModelAndView showCommentForm(@PathVariable("id") Long id, Principal principal) {
 		ModelAndView mv = new ModelAndView();
 		Comment comment = new Comment();
-
 		Optional<Spot> spotOpt = spotRepository.findById(id);
-		comment.setSpot(spotOpt.get());
-				
+		if(spotOpt.isPresent()==false) {
+			mv.setViewName("map");
+			mv.addObject("error", "The spot to be commented does not exist");
+			return mv;
+		}
+
 		mv.setViewName("comment/addComment");
+		comment.setSpot(spotOpt.get());
 		mv.addObject("commentForm", comment);
 		return mv;
 		
 	}
 	
 	@PostMapping("comments/add/save")
-	public String saveComment(@ModelAttribute("commentForm") Comment commentForm, Principal principal) {
+	public String saveComment(@ModelAttribute("commentForm") Comment commentForm, Principal principal, Model model) {
 		Optional<Spot> spotOpt = spotRepository.findById(commentForm.getSpot().getId());
 		commentForm.setSpot(spotOpt.get());
 		commentForm.setDate(LocalDate.now());
@@ -69,10 +74,14 @@ public class CommentsAddController {
 		if(oLoggedUser.isPresent() == false) {
 			System.out.println("Kein User eingeloggt");
 		}
+		if(commentForm.getMessage().isBlank()) {
+			model.addAttribute("error", "Comment Message must not be empty!");
+			return "comment/addComment";
+		}
 		commentForm.setUser(oLoggedUser.get());
 		Comment comment = commentRepository.save(commentForm);
 		System.out.println("Kommentar mit der ID: " + comment.getId() + " gespeichert. Comment Spot: " + comment.getSpot());
-		return("map");
+		return("redirect:/comments/" + comment.getSpot().getId());
 	}
 	
 	@RequestMapping("/comments/{id}")
@@ -82,6 +91,7 @@ public class CommentsAddController {
 		if(spotOpt.isPresent() == false) {
 			System.out.println("Spot " + id + " existiert nicht!");
 			mv.setViewName("map");
+			mv.addObject("error", "This spot does not exist!");
 			return mv;
 		}
 		Spot spot = spotOpt.get();
@@ -90,6 +100,7 @@ public class CommentsAddController {
 		mv.setViewName("comment/showComments");
 		mv.addObject("comments", comments);
 		mv.addObject("userName", principal.getName());
+		mv.addObject("spotId", spot.getId());
 		
 		return mv;
 	}
@@ -97,10 +108,13 @@ public class CommentsAddController {
 	@GetMapping("/comments/edit/{id}")
 	public ModelAndView commentEditForm(@PathVariable("id") Long id, Principal principal) {
 		Optional<Comment> optComment = commentRepository.findById(id);
-		Comment comment = optComment.get();
-		
 		ModelAndView mv = new ModelAndView();
-		
+		if(optComment.isPresent()==false) {
+			mv.setViewName("map");
+			mv.addObject("error", "The comment to be edited does not exist!");
+			return mv;
+		}
+		Comment comment = optComment.get();		
 		Optional<User> oLoggedUser = userRepository.findUserByLogin(principal.getName());
 		
 		if(comment.getUser().getId() != oLoggedUser.get().getId()) {
@@ -123,8 +137,12 @@ public class CommentsAddController {
 	}
 	
 	@GetMapping("/comments/delete/{id}")
-	public String deleteComment(@PathVariable("id") Long id) {
+	public String deleteComment(@PathVariable("id") Long id, Model model) {
 		Optional<Comment> optComment = commentRepository.findById(id);
+		if(optComment.isPresent()==false) {
+			model.addAttribute("error", "The comment to be deleted does not exist!");
+			return "map";
+		}
 		Comment comment = optComment.get();
 		Long spotId = comment.getSpot().getId();
 		commentRepository.delete(comment);
